Move music management code to own package folder
