+++ src/scripts/dump_playlist_genres.py
@@ -3,8 +3,8 @@
 import sys
 sys.path.extend(['.', '../'])
 
-from app.lib.music_util import MusicUtil
-from app.lib.my_music_lib import MyMusicLib
+from packages.music_management.music_util import MusicUtil
+from packages.music_management.my_music_lib import MyMusicLib
 from packages.spotify.spotify_client_wrapper import SpotifyClientWrapper
