+++ src/scripts/update_seed_playlists.py
@@ -3,8 +3,8 @@
 import sys
 sys.path.extend(['.', '../'])
 
-from app.lib.my_music_lib import MyMusicLib
-from app.lib.music_util import MusicUtil
+from packages.music_management.my_music_lib import MyMusicLib
+from packages.music_management.music_util import MusicUtil
 from packages.spotify.spotify_client_wrapper import SpotifyClientWrapper
 
 NUM_TRACKS_PER_ALBUM = 3
